chore: split ViewRoundPage components in different files
