+++ packages/grant-explorer/src/features/round/ViewRoundPage/index.ts
@@ -1,0 +1,4 @@
+import ViewRoundPage from "./ViewRoundPage";
+export * from "./AlloVersionBanner";
+export * from "./CartButtonToggle";
+export default ViewRoundPage;
