+++ packages/grant-explorer/src/features/round/ViewRoundPage/utils.ts
@@ -1,0 +1,6 @@
+export const formatAmount = (amount: string | number, noDigits?: boolean) => {
+  return Number(amount).toLocaleString("en-US", {
+    maximumFractionDigits: noDigits ? 0 : 2,
+    minimumFractionDigits: noDigits ? 0 : 2,
+  });
+};
